requirements.txt
rumps>=0.4.0
requests>=2.31
python-dotenv>=1.0
pyobjc-framework-Cocoa>=10.0
